requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
markdown==3.5.1
weasyprint==60.1
gunicorn==21.2.0

